requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
motor          # async MongoDB driver
passlib[bcrypt]==1.7.4
python-jose[cryptography]
email-validator